+++ examples/data-sources/fmc_icmpv6_objects/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_icmpv6_objects" "example" {
+  items = {
+    "my_icmpv6_objects" = {
+    }
+  }
+}