+++ internal/provider/model_fmc_device_etherchannel_interface.go
@@ -55,6 +55,7 @@ type DeviceEtherChannelInterface struct {
 	NveOnly                               types.Bool                                                  `tfsdk:"nve_only"`
 	Ipv4StaticAddress                     types.String                                                `tfsdk:"ipv4_static_address"`
 	Ipv4StaticNetmask                     types.String                                                `tfsdk:"ipv4_static_netmask"`
+	Ipv4AddressPoolId                     types.String                                                `tfsdk:"ipv4_address_pool_id"`
 	Ipv4DhcpObtainRoute                   types.Bool                                                  `tfsdk:"ipv4_dhcp_obtain_route"`
 	Ipv4DhcpRouteMetric                   types.Int64                                                 `tfsdk:"ipv4_dhcp_route_metric"`
 	Ipv4PppoeVpdnGroupName                types.String                                                `tfsdk:"ipv4_pppoe_vpdn_group_name"`
@@ -69,6 +70,7 @@ type DeviceEtherChannelInterface struct {
 	Ipv6LinkLocalAddress                  types.String                                                `tfsdk:"ipv6_link_local_address"`
 	Ipv6EnableAutoConfig                  types.Bool                                                  `tfsdk:"ipv6_enable_auto_config"`
 	Ipv6Addresses                         []DeviceEtherChannelInterfaceIpv6Addresses                  `tfsdk:"ipv6_addresses"`
+	Ipv6AddressPoolId                     types.String                                                `tfsdk:"ipv6_address_pool_id"`
 	Ipv6Prefixes                          []DeviceEtherChannelInterfaceIpv6Prefixes                   `tfsdk:"ipv6_prefixes"`
 	Ipv6EnableDad                         types.Bool                                                  `tfsdk:"ipv6_enable_dad"`
 	Ipv6DadAttempts                       types.Int64                                                 `tfsdk:"ipv6_dad_attempts"`
@@ -213,6 +215,9 @@ func (data DeviceEtherChannelInterface) toBody(ctx context.Context, state Device
 	if !data.Ipv4StaticNetmask.IsNull() {
 		body, _ = sjson.Set(body, "ipv4.static.netmask", data.Ipv4StaticNetmask.ValueString())
 	}
+	if !data.Ipv4AddressPoolId.IsNull() {
+		body, _ = sjson.Set(body, "ipv4.static.pool.id", data.Ipv4AddressPoolId.ValueString())
+	}
 	if !data.Ipv4DhcpObtainRoute.IsNull() {
 		body, _ = sjson.Set(body, "ipv4.dhcp.enableDefaultRouteDHCP", data.Ipv4DhcpObtainRoute.ValueBool())
 	}
@@ -267,6 +272,9 @@ func (data DeviceEtherChannelInterface) toBody(ctx context.Context, state Device
 			}
 			body, _ = sjson.SetRaw(body, "ipv6.addresses.-1", itemBody)
 		}
+	}
+	if !data.Ipv6AddressPoolId.IsNull() {
+		body, _ = sjson.Set(body, "ipv6.pool.id", data.Ipv6AddressPoolId.ValueString())
 	}
 	if len(data.Ipv6Prefixes) > 0 {
 		body, _ = sjson.Set(body, "ipv6.prefixes", []interface{}{})
@@ -519,6 +527,11 @@ func (data *DeviceEtherChannelInterface) fromBody(ctx context.Context, res gjson
 	} else {
 		data.Ipv4StaticNetmask = types.StringNull()
 	}
+	if value := res.Get("ipv4.static.pool.id"); value.Exists() {
+		data.Ipv4AddressPoolId = types.StringValue(value.String())
+	} else {
+		data.Ipv4AddressPoolId = types.StringNull()
+	}
 	if value := res.Get("ipv4.dhcp.enableDefaultRouteDHCP"); value.Exists() {
 		data.Ipv4DhcpObtainRoute = types.BoolValue(value.Bool())
 	} else {
@@ -607,6 +620,11 @@ func (data *DeviceEtherChannelInterface) fromBody(ctx context.Context, res gjson
 			(*parent).Ipv6Addresses = append((*parent).Ipv6Addresses, data)
 			return true
 		})
+	}
+	if value := res.Get("ipv6.pool.id"); value.Exists() {
+		data.Ipv6AddressPoolId = types.StringValue(value.String())
+	} else {
+		data.Ipv6AddressPoolId = types.StringNull()
 	}
 	if value := res.Get("ipv6.prefixes"); value.Exists() {
 		data.Ipv6Prefixes = make([]DeviceEtherChannelInterfaceIpv6Prefixes, 0)
@@ -979,6 +997,11 @@ func (data *DeviceEtherChannelInterface) fromBodyPartial(ctx context.Context, re
 	} else {
 		data.Ipv4StaticNetmask = types.StringNull()
 	}
+	if value := res.Get("ipv4.static.pool.id"); value.Exists() && !data.Ipv4AddressPoolId.IsNull() {
+		data.Ipv4AddressPoolId = types.StringValue(value.String())
+	} else {
+		data.Ipv4AddressPoolId = types.StringNull()
+	}
 	if value := res.Get("ipv4.dhcp.enableDefaultRouteDHCP"); value.Exists() && !data.Ipv4DhcpObtainRoute.IsNull() {
 		data.Ipv4DhcpObtainRoute = types.BoolValue(value.Bool())
 	} else {
@@ -1096,6 +1119,11 @@ func (data *DeviceEtherChannelInterface) fromBodyPartial(ctx context.Context, re
 			data.EnforceEui = types.BoolNull()
 		}
 		(*parent).Ipv6Addresses[i] = data
+	}
+	if value := res.Get("ipv6.pool.id"); value.Exists() && !data.Ipv6AddressPoolId.IsNull() {
+		data.Ipv6AddressPoolId = types.StringValue(value.String())
+	} else {
+		data.Ipv6AddressPoolId = types.StringNull()
 	}
 	for i := 0; i < len(data.Ipv6Prefixes); i++ {
 		keys := [...]string{"address", "default"}