+++ internal/provider/resource_fmc_device_subinterface.go
@@ -185,6 +185,10 @@ func (r *DeviceSubinterfaceResource) Schema(ctx context.Context, req resource.Sc
 				MarkdownDescription: helpers.NewAttributeDescription("Netmask (width) for ipv4_static_address.").String,
 				Optional:            true,
 			},
+			"ipv4_address_pool_id": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Id of the assigned IPv4 address pool.").String,
+				Optional:            true,
+			},
 			"ipv4_dhcp_obtain_route": schema.BoolAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("Any non-null value here indicates to enable DHCPv4. Value `false` indicates to enable DHCPv4 without obtaining from there the default IPv4 route but anyway requires also ipv4_dhcp_route_metric to be set to exactly 1. Value `true` indicates to enable DHCPv4 and obtain the route and also requires ipv4_dhcp_route_metric to be non-null. The ipv4_dhcp_obtain_route must be null when using ipv4_static_address.").String,
 				Optional:            true,
@@ -265,6 +269,10 @@ func (r *DeviceSubinterfaceResource) Schema(ctx context.Context, req resource.Sc
 						},
 					},
 				},
+			},
+			"ipv6_address_pool_id": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Id of the assigned IPv6 address pool.").String,
+				Optional:            true,
 			},
 			"ipv6_prefixes": schema.ListNestedAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("").String,